Produce error messages where the error first occurs.

+++ gdk/gdk_bbp.c
@@ -1286,23 +1286,19 @@ new_bbpentry(FILE *fp, bat i)
 		    (unsigned char) BBP_desc(i)->S.map_head,
 		    (unsigned char) BBP_desc(i)->S.map_tail,
 		    (unsigned char) BBP_desc(i)->S.map_hheap,
-		    (unsigned char) BBP_desc(i)->S.map_theap) < 0)
+		    (unsigned char) BBP_desc(i)->S.map_theap) < 0 ||
+	    heap_entry(fp, &BBP_desc(i)->H) < 0 ||
+	    heap_entry(fp, &BBP_desc(i)->T) < 0 ||
+	    vheap_entry(fp, BBP_desc(i)->H.vheap) < 0 ||
+	    vheap_entry(fp, BBP_desc(i)->T.vheap) < 0 ||
+	    (BBP_options(i) &&
+	     fprintf(fp, " %s", BBP_options(i)) < 0) ||
+	    fprintf(fp, "\n") < 0) {
+		GDKsyserror("new_bbpentry: Writing BBP.dir entry failed\n");
 		return GDK_FAIL;
-	if (heap_entry(fp, &BBP_desc(i)->H) < 0)
-		return GDK_FAIL;
-	if (heap_entry(fp, &BBP_desc(i)->T) < 0)
-		return GDK_FAIL;
+	}
 
-	if (vheap_entry(fp, BBP_desc(i)->H.vheap) < 0)
-		return GDK_FAIL;
-	if (vheap_entry(fp, BBP_desc(i)->T.vheap) < 0)
-		return GDK_FAIL;
-
-	if (BBP_options(i) &&
-	    fprintf(fp, " %s", BBP_options(i)) < 0)
-		return GDK_FAIL;
-
-	return fprintf(fp, "\n") < 0 ? GDK_FAIL : GDK_SUCCEED;
+	return GDK_SUCCEED;
 }
 
 static gdk_return
@@ -1312,8 +1308,10 @@ BBPdir_header(FILE *f, int n)
 		    GDKLIBRARY, SIZEOF_SIZE_T, SIZEOF_OID, SIZEOF_MAX_INT) < 0 ||
 	    OIDwrite(f) < 0 ||
 	    fprintf(f, " BBPsize=%d\n", n) < 0 ||
-	    ferror(f))
+	    ferror(f)) {
+		GDKsyserror("BBPdir_header: Writing BBP.dir header failed\n");
 		return GDK_FAIL;
+	}
 	return GDK_SUCCEED;
 }
 
@@ -1328,10 +1326,8 @@ BBPdir_subcommit(int cnt, bat *subcommit)
 
 	assert(subcommit != NULL);
 
-	if ((nbbpf = GDKfilelocate(0, "BBP", "w", "dir")) == NULL) {
-		GDKsyserror("BBPdir_subcommit: Creating new BBP.dir file failed\n");
+	if ((nbbpf = GDKfilelocate(0, "BBP", "w", "dir")) == NULL)
 		return GDK_FAIL;
-	}
 
 	n = (bat) ATOMIC_GET(BBPsize, BBPsizeLock, "BBPdir_subcommit");
 
@@ -1361,7 +1357,6 @@ BBPdir_subcommit(int cnt, bat *subcommit)
 	}
 
 	if (BBPdir_header(nbbpf, n) != GDK_SUCCEED) {
-		GDKsyserror("BBPdir_subcommit: Writing BBP.dir header failed\n");
 		goto bailout;
 	}
 	n = 0;
@@ -1385,7 +1380,6 @@ BBPdir_subcommit(int cnt, bat *subcommit)
 			/* BBP.dir consists of all persistent bats only */
 			if (BBP_status(i) & BBPPERSISTENT) {
 				if (new_bbpentry(nbbpf, i) != GDK_SUCCEED) {
-					GDKsyserror("BBPdir_subcommit: Writing BBP.dir entry failed\n");
 					goto bailout;
 				}
 				IODEBUG new_bbpentry(stderr, i);
@@ -1457,12 +1451,10 @@ BBPdir(int cnt, bat *subcommit)
 		fprintf(stderr, "\n");
 	}
 	if ((fp = GDKfilelocate(0, "BBP", "w", "dir")) == NULL) {
-		GDKsyserror("BBPdir: Cannor create BBP.dir file\n");
 		goto bailout;
 	}
 
 	if (BBPdir_header(fp, (bat) ATOMIC_GET(BBPsize, BBPsizeLock, "BBPdir")) != GDK_SUCCEED) {
-		GDKsyserror("BBPdir: Writing BBP.dir header failed\n");
 		goto bailout;
 	}
 
@@ -1471,7 +1463,6 @@ BBPdir(int cnt, bat *subcommit)
 		 * BBP.dir consists of all persistent bats */
 		if (BBP_status(i) & BBPPERSISTENT) {
 			if (new_bbpentry(fp, i) != GDK_SUCCEED) {
-				GDKsyserror("BBPdir: Writing BBP.dir entry failed\n");
 				goto bailout;
 			}
 			IODEBUG new_bbpentry(stderr, i);
@@ -3175,12 +3166,15 @@ file_move(int farmid, const char *srcdir, const char *dstdir, const char *name, 
 		struct stat st;
 
 		path = GDKfilepath(farmid, srcdir, name, ext);
+		if (path == NULL)
+			return GDK_FAIL;
 		if (stat(path, &st)) {
 			/* source file does not exist; the best
 			 * recovery is to give an error but continue
 			 * by considering the BAT as not saved; making
 			 * sure that this time it does get saved.
 			 */
+			GDKsyserror("file_move: cannot stat %s\n", path);
 			GDKfree(path);
 			return GDK_FAIL;	/* fishy, but not fatal */
 		}
@@ -3195,12 +3189,14 @@ file_exists(int farmid, const char *dir, const char *name, const char *ext)
 {
 	char *path;
 	struct stat st;
-	int ret;
+	int ret = -1;
 
 	path = GDKfilepath(farmid, dir, name, ext);
-	ret = stat(path, &st);
-	IODEBUG fprintf(stderr, "#stat(%s) = %d\n", path, ret);
-	GDKfree(path);
+	if (path) {
+		ret = stat(path, &st);
+		IODEBUG fprintf(stderr, "#stat(%s) = %d\n", path, ret);
+		GDKfree(path);
+	}
 	return (ret == 0);
 }
 
@@ -3231,7 +3227,11 @@ heap_move(Heap *hp, const char *srcdir, const char *dstdir, const char *nme, con
 
 		snprintf(kill_ext, sizeof(kill_ext), "%s.kill", ext);
 		path = GDKfilepath(hp->farmid, dstdir, nme, kill_ext);
+		if (path == NULL)
+			return GDK_FAIL;
 		fp = fopen(path, "w");
+		if (fp == NULL)
+			GDKsyserror("heap_move: cannot open file %s\n", path);
 		IODEBUG fprintf(stderr, "#open %s = %d\n", path, fp ? 0 : -1);
 		GDKfree(path);
 
@@ -3280,16 +3280,20 @@ BBPprepare(bit subcommit)
 		backup_dir = 0;
 		ret = BBPrecover(0);
 		if (ret == GDK_SUCCEED) {
-			if (mkdir(BAKDIR, 0755) < 0 && errno != EEXIST)
+			if (mkdir(BAKDIR, 0755) < 0 && errno != EEXIST) {
+				GDKsyserror("BBPprepare: cannot create directory %s\n", BAKDIR);
 				ret = GDK_FAIL;
+			}
 			/* if BAKDIR already exists, don't signal error */
 			IODEBUG fprintf(stderr, "#mkdir %s = %d\n", BAKDIR, (int) ret);
 		}
 	}
 	if (ret == GDK_SUCCEED && start_subcommit) {
 		/* make a new SUBDIR (subdir of BAKDIR) */
-		if (mkdir(SUBDIR, 0755) < 0)
+		if (mkdir(SUBDIR, 0755) < 0) {
+			GDKsyserror("BBPprepare: cannot create directory %s\n", SUBDIR);
 			ret = GDK_FAIL;
+		}
 		IODEBUG fprintf(stderr, "#mkdir %s = %d\n", SUBDIR, (int) ret);
 	}
 	if (ret == GDK_SUCCEED && backup_dir != set) {
@@ -3594,8 +3598,6 @@ force_move(int farmid, const char *srcdir, const char *dstdir, const char *name)
 
 		(void) GDKcreatedir(dstdir); /* if fails, move will fail */
 		ret = GDKmove(farmid, srcdir, name, NULL, dstdir, name, NULL);
-		if (ret != GDK_SUCCEED)
-			GDKsyserror("force_move: link(%s,%s)=%d\n", srcpath, dstpath, (int) ret);
 		IODEBUG fprintf(stderr, "#link %s %s = %d\n", srcpath, dstpath, (int) ret);
 		GDKfree(dstpath);
 		GDKfree(srcpath);
@@ -3625,6 +3627,7 @@ BBPrecover(int farmid)
 	IODEBUG fprintf(stderr, "#BBPrecover(start)\n");
 
 	if (mkdir(LEFTDIR, 0755) < 0 && errno != EEXIST) {
+		GDKsyserror("BBPrecover: cannot create directory %s\n", LEFTDIR);
 		closedir(dirp);
 		return GDK_FAIL;
 	}
@@ -3634,17 +3637,18 @@ BBPrecover(int farmid)
 		const char *q = strchr(dent->d_name, '.');
 
 		if (q == dent->d_name) {
-			int uret;
 			char *fn;
 
 			if (strcmp(dent->d_name, ".") == 0 ||
 			    strcmp(dent->d_name, "..") == 0)
 				continue;
 			fn = GDKfilepath(farmid, BAKDIR, dent->d_name, NULL);
-			uret = unlink(fn);
-			IODEBUG fprintf(stderr, "#unlink %s = %d\n", fn, uret);
-			GDKfree(fn);
-
+			if (fn) {
+				int uret = unlink(fn);
+				IODEBUG fprintf(stderr, "#unlink %s = %d\n",
+						fn, uret);
+				GDKfree(fn);
+			}
 			continue;
 		} else if (strcmp(dent->d_name, "BBP.dir") == 0) {
 			dirseen = TRUE;
@@ -3684,8 +3688,10 @@ BBPrecover(int farmid)
 	}
 
 	if (ret == GDK_SUCCEED) {
-		if (rmdir(BAKDIR) < 0)
+		if (rmdir(BAKDIR) < 0) {
+			GDKsyserror("BBPrecover: cannot remove directory %s\n", BAKDIR);
 			ret = GDK_FAIL;
+		}
 		IODEBUG fprintf(stderr, "#rmdir %s = %d\n", BAKDIR, (int) ret);
 	}
 	if (ret != GDK_SUCCEED)
